+++ packages/sdk/src/sdk.ts
@@ -43,6 +43,10 @@ export class ArenaChat {
   private unsubscribeOnUnreadMessagesCountChanged: (() => void) | undefined;
   private liveChat: LiveChat | null = null;
   private userProfileI: BaseUserProfile | null = null;
+  private promiseFetchAndSetChatRoomAndSite: Promise<{
+    chatRoom: ChatRoom;
+    site: Site;
+  }> | null = null;
 
   public constructor(private apiKey: string) {
     this.restAPI = new RestAPI({ authToken: this.defaultAuthToken });
@@ -234,7 +238,11 @@ export class ArenaChat {
     }
 
     try {
-      const { chatRoom, site } = await this.fetchAndSetChatRoomAndSite(slug);
+      if (this.promiseFetchAndSetChatRoomAndSite === null) {
+        this.promiseFetchAndSetChatRoomAndSite = this.fetchAndSetChatRoomAndSite(slug);
+      }
+
+      const { chatRoom, site } = await this.promiseFetchAndSetChatRoomAndSite;
 
       const liveChatI = new LiveChat(chatRoom, site, this);
 
@@ -342,7 +350,7 @@ export class ArenaChat {
    *
    * @param user created user
    */
-  private callChangedUserListeners(user: ExternalUser): void {
+  private callChangedUserListeners(user: ExternalUser | null): void {
     this.userChangedListeners.forEach((listener) => {
       listener(user);
     });
@@ -355,6 +363,7 @@ export class ArenaChat {
   private unsetUser(): void {
     this.user = null;
     this.restAPI = new RestAPI({ authToken: this.defaultAuthToken });
+    this.callChangedUserListeners(null);
   }
 
   /**