+++ packages/sdk/src/services/graphql-api.ts
@@ -57,6 +57,8 @@ export class GraphQLAPI {
               image
               isBlocked
               status
+              isModerator
+              modLabel
             }
             name
             unreadCount
@@ -140,6 +142,8 @@ export class GraphQLAPI {
             image
             isBlocked
             status
+            isModerator
+            modLabel
           }
           name
           amIBlocked
@@ -170,6 +174,7 @@ export class GraphQLAPI {
               status
               modLabel
               isModerator
+              modLabel
               image
               isBlocked
             }
@@ -215,6 +220,8 @@ export class GraphQLAPI {
             image
             isBlocked
             status
+            isModerator
+            modLabel
           }
           name
           unreadCount